backend/requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
pydantic~=2.11.3
sqlalchemy~=2.0.40
asyncpg
pandas~=2.2.3

passlib~=1.7.4
python-jose~=3.5.0